Some modifications to snipMate
Added a "wildcard" snippet -- if no named snippet is found, the wildcard is
expanded, and a token is replaced by the nearest word. Could be very useful for
html and eventually, maybe for LaTeX (I should be needing that soon enough...).

Added the showmarks plugin, which opens up an entire new world of
possibilities, but it still needs some tweaking to be usable.

Also, updates for a few plugins.