Supertab
Supertab is a great plugin, but it's a sore reminder that I need to work
on the PHP omnicompletion. Currently, the strategy is to load everything
the first time any completion is requested and then query it in the
memory. That doesn't seem like an optimal solution, because even in
memory, the search is still slow-ish, and the load time at the beginning
is ridiculous. Complete rewrite seems the only adequate solution in my
case -- I just have to make incremental searches and cache it as I go. A
better data structure is a must.  Possible integration with smarttab
would be a bonus.

Moved all functions from ghc.vim to autoload. Speeds it up a
bit... Maybe.

Removed the 'set path=./**', vim hangs sometimes (apparently, it's a
known bug...)