Completion is a bitch.
Evidently, getting proper completion to work is not as easy as expected (just
like anything else, I suppose). Supertab turned out to be not as useful as
expected, so that's removed as well.

Good news is that there is a new version of NERDTree that plays a lot better
with my small Utl function (now officially dubbed "plugin"). Still, it's a bit
ironic that a plugin for vim can now have plugins of its own...