FuzzyFinderTextMate working
Needed to downgrade ruby, but I haven't really been using 1.9 much, so
it's not much of a problem. And thankfully, the windows binary comes
with dynamic ruby loading, so it won't be needing recompiling.

Removed FindFile -- pretty useless, when you have the FuzzyFinder.
Still, if the latter turns out to be too slow in actual projects, it
might be a better idea to go back to FindFile...

Removed Abolish for now -- don't have any motivation to use it.