GTD
Started trying to implement the GTD methodology again. Unfortunately,
the AGTD plugin is still tricky to set up and not very pretty on the
terminal. Add to that a few small items I'm not comfortable with and it
seems it's best for me to think about my own solution to this.
'syntax/gtd.vim' is the start for that, and agtd-related stuff are going
to serve as example.