GTD: changed to SGTD
SGTD is supposed to stand for "Simple GTD", and let's hope it really
does. For now, I added a syntax group for anything that looks like a
"date: ....". The reason for not sticking to a specific implementation
is that it might be useful to write not a date, but a string like
"Tomorrow" and think of a good system to manage such time descriptions.