Snippet cleanup
Deleted a few snippet sets that I never used (if I need snips for perl
or java, I can always download the plugin again), tweaked a few other
ones, fixed a few bugs.

There are still a lot of unused snips, but I'm still hoping they might
be of use in the future.