Rake completion taken from rails.vim
Caching is not very intelligent -- the rake tasks are cached per buffer.
Still, in a rails project, the rails.vim Rake command will take over and
in other projects, it's unlikely that there will be so many rake tasks
to begin with (so that might mean it's best to just ignore caching
altogether, but let's see what issues arise first).