Mapping to highlight grouped items with matchit
The normal '%' mapping provided by the matchit plugin is very useful,
but sometimes not clear enough in long bodies of code. It's not very
complicated to use it to highlight all the items it would go through,
allowing a better view of if/then/else statements, functions with their
return statements, etc.