NERDTree plugin to follow currently open file
As cautious as possible, only happens if the tree is open, ignores
errors for new files that are not yet updated, stuff like that. Has
commands to turn it off and on at will.

For now, using just for vimfiles directory, to see if it works well,
might enable it globally.